Styled the Navigation
Styled the Navigation using  React Bootstrap.
