+++ package.json
@@ -25,11 +25,13 @@
   "homepage": "https://github.com/myfriendpaul/boom-roasted#readme",
   "dependencies": {
     "bcrypt": "^5.0.1",
+    "bootstrap": "^5.1.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.5",
-    "morgan": "^1.10.0"
+    "morgan": "^1.10.0",
+    "react-bootstrap": "^2.0.0-beta.6"
   },
   "type": "module",
   "devDependencies": {
